+++ .rsink/constants.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
-version="rsink 0.2"
-config_file="config"
-profiles_directory="profiles"
-tools_directory="tools"
-log_file="log"
+VERSION="rsink 0.2+dev"
+CONFIG_FILE="config"
+PROFILES_DIRECTORY="profiles"
+TOOLS_DIRECTORY="tools"
+LOG_FILE="log"